improving profile image upload
