+++ src/services/axios-config.js
@@ -18,4 +18,3 @@ instance.interceptors.request.use(
 );
 
 export default instance
-
